Iheatmap-class

Class to store complex interactive heatmap objects from iheatmapr package

Details

This is a virtual class with two children classes, IheatmapHorizontal and IheatmapVertical. For IheatmapHorizontal additional main heatmaps are added horizontally, and for IheatmapVertical additional main heatmaps are added vertically. For details on accessing certain slots of this class, see access_component documentation.

Slots

plots

list of plot element in IheatmapPlots format

shapes

list of shape element in IheatmapShapes format

annotations

list of annotation elements in IheatmapAnnotations format

xaxes

list of x axes in IheatmapAxes format

yaxes

list of y axes in IheatmapAxes format

colorbars

list of colorbars in IheatmapColorbars format

colorbar_grid

colorbar grid parameters in IheatmapColorbarGrid format

current_xaxis

name of current x axis

current_yaxis

name of current y axis

layout

list of plotly layout parameters

See also

iheatmap, main_heatmap, access_component