add_row_labels

Add y axis labels to plot

# S4 method for Iheatmap
add_row_labels(p, tickvals = NULL, ticktext = NULL,
  textangle = 0, font = list(), side = c("left", "right"), size = 0.1,
  buffer = 0.005, xname = NULL, yname = current_yaxis(p))

Arguments

p

Iheatmap-class object

tickvals

row indices at which to place axis tick labels

ticktext

text for axis tick labels

textangle

angle for ticktext

font

list of plotly font attributes, see https://plot.ly/javascript/reference/#layout-font

side

side of plot on which to add subplot

size

relative size of subplot relative to main heatmap

buffer

amount of space to leave empty before this plot, relative to size of first heatmap

xname

internal name for xaxis

yname

internal name for yaxis

Value

Iheatmap-class object, which can be printed to generate an interactive graphic

See also

add_row_title, iheatmap, add_col_labels

Examples

mat <- matrix(rnorm(20), ncol = 5, nrow = 4)
iheatmap(mat) %>% add_row_labels()
iheatmap(mat) %>% add_row_labels(ticktext = letters[23:26])